% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractCalibrationPoints.R
\name{extractCalibrationPoints}
\alias{extractCalibrationPoints}
\title{extractCalibrationPoints}
\usage{
extractCalibrationPoints(type = "p", pch = 4, col = "blue", lwd = 2,
  roundValue = 5)
}
\arguments{
\item{type}{same as pch in bas graphics}

\item{pch}{same as pch base graphics}

\item{col}{same as col in base graphics}

\item{lwd}{same as lwd in base graphics}

\item{roundValue}{number of decimal points to round text output}
}
\description{
This function extracts calibration points.
}
\details{
The text printed from this function can be copied into R code.
This way the image does not need to be recalibrated every time the code is run.
}
\examples{


prepareExtractionPlot("figure_7_Rogers_Commission_1986.png")
calpoints <- extractCalibrationPoints()


}
\keyword{time}
