% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareExtractionPlot.R
\name{prepareExtractionPlot}
\alias{prepareExtractionPlot}
\title{prepareExtractionPlot.R}
\usage{
prepareExtractionPlot(file, workingDirectory = NULL,
  main = "Data Extraction Plot", xlab = "x", ylab = "y")
}
\arguments{
\item{file}{png file of data to be extracted}

\item{workingDirectory}{the directory the file is located. If this value is
null the function assumes it is in the working directory}

\item{main}{main label for the plots}

\item{xlab}{label for the x axis}

\item{ylab}{label for the y axis}
}
\description{
This function plots an image to a graph.  This is necissary to extract the data.
}
\examples{


prepareExtractionPlot("figure_7_Rogers_Commission_1986.png")

}
\keyword{extraction}
\keyword{plot}
\keyword{point}
