% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recoverDataFromDatePoints.R,
%   R/recoverDataFromPoints.R
\name{recoverDataFromDatePoints}
\alias{recoverDataFromDatePoints}
\alias{recoverDataFromDatePoints}
\title{recoverDataFromDatePoints}
\usage{
recoverDataFromDatePoints(calibrationPoints, dataPoints, xmin, xmax, ymin, ymax,
  dateFormat = "\%m/\%d/\%Y", roundValue = 2)

recoverDataFromDatePoints(calibrationPoints, dataPoints, xmin, xmax, ymin, ymax,
  dateFormat = "\%m/\%d/\%Y", roundValue = 2)
}
\arguments{
\item{calibrationPoints}{a list of calibration points as
a list of x coordinates and another list with y coordinates.
The points are in the order of ymax, ymin, xmax, xmin}

\item{dataPoints}{a list of data points as
a list of x coordinates and another list with y coordinates.}

\item{xmin}{The minimum on the x axis of an extraction plot}

\item{xmax}{The maximum on the x axis of an extraction plot}

\item{ymin}{The minimum on the y axis of an extraction plot}

\item{ymax}{The maximum on the y axis of an extraction plot}

\item{dateFormat}{The formatting for the date.  The default is "%m/%d/%Y".
The formatting information for the date is the same as the as.Date function.}

\item{roundValue}{The number of decimal points to round data.}

\item{beginDate}{The beginning of a time coordinate on the x axis}

\item{endDate}{The end of a time coordinate on the x axis}

\item{calibrationPoints}{a list of calibration points as
a list of x coordinates and another list with y coordinates.
The points are in the order of ymax, ymin, xmax, xmin}

\item{dataPoints}{a list of data points as
a list of x coordinates and another list with y coordinates.}

\item{ymin}{The minimum on the y axis of an extraction plot}

\item{ymax}{The maximum on the y axis of an extraction plot}

\item{dateFormat}{The formatting for the date.  The default is "%m/%d/%Y".
The formatting information for the date is the same as the as.Date function.}

\item{roundValue}{The number of decimal points to round data.}
}
\description{
The function recovers data from a plot given the calibration points and
data points.

The function recovers data from a plot given the calibration points and
data points.
}
\details{
It is assumed the figure is rotated such that the x and y axis are in the
cardinal directions.  A distorted or rotated image should be processed with
image editing software before a recovery is attempted.

The extractCalibrationPoints() function can aquire the calibration points.
The extractDataPoints() function can aquire the data points.

This function is designed to work on a single dataset.  If multiple datasets are
in the software then this function should be called multiple times.

This function assumes time progresses from left to right on the x axis.
This function assumes the value goes from the minimum on the bottom to the
maximum on the top.

It is assumed the figure is rotated such that the x and y axis are in the
cardinal directions.  A distorted or rotated image should be processed with
image editing software before a recovery is attempted.

The extractCalibrationPoints() function can aquire the calibration points.
The extractDataPoints() function can aquire the data points.

This function is designed to work on a single dataset.  If multiple datasets are
in the software then this function should be called multiple times.

This function assumes min is on left and max is on right.
This function assumes the value goes from the minimum on the bottom to the
maximum on the top.
}
\examples{

prepareExtractionPlot(
   "figure_1_Global_Temperature_Change_Hansen_Et_Al_2006.png"
   )

calibrationPoints <- extractCalibrationPoints()
datapoints <- extractDataPoints(24)




prepareExtractionPlot("figure_7_Rogers_Commission_1986.png")
calibrationPoints <- extractCalibrationPoints()
datapoints <- extractDataPoints(24)

dataValues <- recoverDataFromDatePoints(calibrationPoints,
   datapoints, beginDate, endDate, ymin, ymax
)


}
\keyword{time}
