% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showCalibrationPoints.R
\name{showCalibrationPoints}
\alias{showCalibrationPoints}
\title{showCalibrationPoints}
\usage{
showCalibrationPoints(calibrationPoints, pch = 4, cex = 2, col = "red",
  lwd = 3)
}
\arguments{
\item{calibrationPoints}{a list with a list of x coordinates and another list
with y coordinates}

\item{pch}{same as pch base graphics}

\item{cex}{same as cex in base graphics}

\item{col}{same as col in base graphics}

\item{lwd}{same as lwd in base graphics}
}
\description{
This is a wrapper function for displaying a set of calibration points on a plot.
}
\examples{

prepareExtractionPlot("figure_7_Rogers_Commission_1986.png")
calpoints <- extractCalibrationPoints()

prepareExtractionPlot("figure_7_Rogers_Commission_1986.png")
showCalibrationPoints(calpoints)

}
\keyword{time}
