% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showDataPoints.R
\name{showDataPoints}
\alias{showDataPoints}
\title{showDataPoints}
\usage{
showDataPoints(dataPoints, dataSetNumber = 1, numofDatasets = 10, pch = 4,
  cex = 2, colorPalette = NULL, lwd = 3)
}
\arguments{
\item{dataPoints}{a list with a list of x coordinates and another list
with y coordinates}

\item{dataSetNumber}{if more than one dataset is included this changes the
color of the dataset}

\item{numofDatasets}{number of datasets to be plotted (only needs to be changed
if more than 10 datasets are being plotted)}

\item{pch}{same as pch base graphics}

\item{lwd}{same as lwd in base graphics}

\item{col}{same as col in base graphics}
}
\description{
This is a wrapper function for displaying a set of
data points on a plot.
}
\examples{

prepareExtractionPlot("figure_7_Rogers_Commission_1986.png")
datapoints <- extractDataPoints()

prepareExtractionPlot("figure_7_Rogers_Commission_1986.png")
showDataPoints(calpoints)

}
\keyword{time}
