% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractDataPoints.R
\name{extractDataPoints}
\alias{extractDataPoints}
\title{extractDataPoints}
\usage{
extractDataPoints(numberOfPoints, numberOfPointsInRow = 7, type = "p",
  pch = 4, col = "blue", lwd = 2, roundValue = 5)
}
\arguments{
\item{numberOfPoints}{the number of datapoints to extract}

\item{numberOfPointsInRow}{the number of points printed per row in output}

\item{type}{same as pch in bas graphics}

\item{pch}{same as pch base graphics}

\item{col}{same as col in base graphics}

\item{lwd}{same as lwd in base graphics}

\item{roundValue}{number of decimal points to round text output}
}
\description{
This function extracts calibration points.
}
\details{
To use this function first prepare an extraction plot using the
prepareExtractionPlot function.

Count the number of points in the dataset to be extracted.

Run the function

Click on every point in the dataset

The text printed from this function can be copied directly as R code.
This way the image does not need to be recalibrated every time the code 
is run.
}
\examples{

prepareExtractionPlot("figure_7_Rogers_Commission_1986.png")
dataPoints <- extractDataPoints(24)

}
\keyword{datapoint}
\keyword{extraction}
